refactor&config/id change from long to Integer & config model mapper and date time format

+++ src/main/java/com/example/backend_rw/entity/Video.java
@@ -14,7 +14,7 @@ public class Video {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
 
     @Column(nullable = false, length = 150)
     private String url;
